pyproject.toml
[tool.poetry]
name = "plex"
version = "0.1.0"
description = "An AI workflow that extracts consolidated profit & loss from financial statements"
readme = "README.md"
authors = ["Ishara Dissanayake <thisismaduishara@gmail.com>"]
packages = [{ include = "plex" }]
classifiers = [
    "Private :: Do Not Upload",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

[tool.poetry.scripts]
plex = "plex:run"

[tool.poetry.dependencies]
python = ">=3.11.9,<3.12"
python-dotenv = "^1.0.1"
click = "^8.1.8"
sanic = "^24.12.0"
sanic-ext = "^23.12.0"
langchain-openai = "^0.1.23"
markitdown = "^0.0.1a3"
motor = "^3.6.1"
scikit-learn = "^1.6.1"
pandas = "^2.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
